app2/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
Pillow==10.0.1
